+++ src/lib/next-button.directive.ts
@@ -9,7 +9,7 @@ import { StepperComponent } from './stepper/stepper.component';
 	selector: 'button[nextButton], button[continueButton]',
 	standalone: true,
 	host: {
-		class: 'stepper__button stepper__button--next',
+		class: 'hub-stepper__button hub-stepper__button--next stepper__button stepper__button--next',
 		'[disabled]': 'disabled()',
 		'(click)': 'stepper.goToNext()'
 	}